requirements.txt
streamlit>=1.26
numpy>=1.23
pillow>=9.5
torch>=2.0
torchvision>=0.15
timm>=0.9.2
requests>=2.31
